fixed encoders

+++ emb-service/processing/models/audio_ast.py
@@ -17,7 +17,7 @@ class AudioASTEncoder(Encoder):
 
         self.model = ASTForAudioClassification.from_pretrained(
             "MIT/ast-finetuned-audioset-10-10-0.4593",
-            torch_dtype=torch.bfloat16,
+            torch_dtype=torch.float16,
             output_hidden_states=True
         ).to(self.device)
         self.feature_extractor = ASTFeatureExtractor.from_pretrained("MIT/ast-finetuned-audioset-10-10-0.4593")
@@ -48,7 +48,7 @@ class AudioASTEncoder(Encoder):
                     return_tensors="pt"
                 )['input_values']
                 for elem in batch
-            ]).to(torch.bfloat16).to(self.device)
+            ]).to(torch.float16).to(self.device)
             outputs = self.model(input_values=inputs)
             cls_embeddings = outputs.hidden_states[-1][:, 0, :]
             embs.append(cls_embeddings.cpu().numpy())
